refactor(ui): share composeRefs from the ui package
TabsList merges its own list ref with the caller's, which is exactly
what composeRefs already does correctly. Move it out of the web app so
the design system can use it.

Merging refs inline is the trap this avoids: forwarding the caller's
return value makes React read whatever their callback happened to
return as a cleanup, and warn when it is neither a function nor
undefined.
